+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.6.23
+- Carrier integration
+  - Added '+' button add steps
+  - Fix:
+    - Fixed module-step not updating correctly in back-end when changing module
+    - Now hides scenario grid when no scenario modular elements present
+- Modular scenarios
+  - Added '+' button to add scenarios
+- Get/set parameters
+  - Fix: 
+    - Fixed outline of 'update all change reasons' field
+
 ## 1.6.22
 - Get/set parameters
   - When creating new parameter codes on PROD, checks if exists in ACC, and if so, pre-fills with description and explanation from ACC